% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v1_0_core_layer.R
\name{auto_recluster}
\alias{auto_recluster}
\title{Auto-reclustering by resolution gradient + ROGUE}
\usage{
auto_recluster(
  seurat_obj,
  resolutions = seq(0.2, 1.6, by = 0.2),
  min_rogue = 0.85
)
}
\description{
Auto-reclustering by resolution gradient + ROGUE
}
