% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v1_1_confidence_layer.R
\name{conformal_predict}
\alias{conformal_predict}
\title{Produce conformal prediction set (A.3)}
\usage{
conformal_predict(prob_test, calib)
}
\arguments{
\item{prob_test}{Numeric matrix (cells x types) of softmax probabilities.}

\item{calib}{Output of \code{split_conformal_calibrate()}.}
}
\value{
List with \code{set} (list of integer vectors), \code{set_size} (integer).
}
\description{
Produce conformal prediction set (A.3)
}
