% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v1_1_confidence_layer.R
\name{hierarchical_reject_one}
\alias{hierarchical_reject_one}
\title{Apply reject option to one cell (A.2 helper)}
\usage{
hierarchical_reject_one(probs, threshold)
}
\description{
Apply reject option to one cell (A.2 helper)
}
