% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v1_0_core_layer.R
\name{merge_similar_numbered_celltypes}
\alias{merge_similar_numbered_celltypes}
\title{Merge similar numbered cell-type clusters by marker Jaccard}
\usage{
merge_similar_numbered_celltypes(
  seurat_obj,
  markers_df,
  jaccard_threshold = 0.8
)
}
\description{
Merge similar numbered cell-type clusters by marker Jaccard
}
