% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v1_0_core_layer.R
\name{normalize_celltype}
\alias{normalize_celltype}
\title{Normalize a free-text cell type to scONCO canonical name}
\usage{
normalize_celltype(x)
}
\description{
Normalize a free-text cell type to scONCO canonical name
}
