% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v1_0_core_layer.R
\name{plot_marker_dotplots}
\alias{plot_marker_dotplots}
\title{Per-cell-type DotPlot panel}
\usage{
plot_marker_dotplots(seurat_obj, markers_df, group_by = "scONCO_L4")
}
\description{
Per-cell-type DotPlot panel
}
