% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v1_0_core_layer.R
\name{sconco_activity_scores}
\alias{sconco_activity_scores}
\title{Cell-type activity scores}
\usage{
sconco_activity_scores(expr, markers_df, delta = 0.2)
}
\description{
score(cell, type) = sum_{g in M+(t)} weight(g) * expr(g, cell)
- delta * sum_{g in M-(t)} expr(g, cell)
}
