% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v1_0_core_layer.R
\name{sconco_annotate}
\alias{sconco_annotate}
\title{Annotate cells directly from a score matrix}
\usage{
sconco_annotate(score_matrix, hierarchy_table = NULL)
}
\description{
Annotate cells directly from a score matrix
}
