% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v1_1_confidence_layer.R
\name{update_specificity_empirical}
\alias{update_specificity_empirical}
\title{Update specificity column empirically (A.1)}
\usage{
update_specificity_empirical(markers_df, expr, labels)
}
\arguments{
\item{markers_df}{Marker data frame.}

\item{expr}{Reference expression matrix.}

\item{labels}{Reference cell-type labels.}
}
\value{
Updated \code{markers_df}.
}
\description{
Recomputes \code{markers_df$specificity_index} using AUC on a reference atlas.
}
